making moves to streamline arg parsing and add in --version
